tidied up other repressilator example components
