+++ examples/repressilator/improved.jl
@@ -1,117 +1,63 @@
-using Catalyst
-using OrdinaryDiffEq
-using Plots
-using JumpProcesses
-import Distributions: Geometric
-using Polynomials
-using Random
+module Improved
+    using Catalyst
+    using StochasticDiffEq
+    import Distributions: Geometric
+    using Polynomials
 
-# Constants
-λ = 60
-K = [5, 10, 10]
-n = 2
-b̄ = 10
-N̄ₒ = 10
-N̄ₜ = 40
+    # Constants
+    λ = 60
+    K = [5, 10, 10]
+    n = 2
+    b̄ = 10
+    N̄ₒ = 10
+    N̄ₜ = 40
 
-@register_symbolic Geometric(b)
-@parameters b
-m = rand(Geometric(1/b)) + 1
+    @register_symbolic Geometric(b)
+    @parameters b
+    m = rand(Geometric(1/b)) + 1
 
-function root(Pt, Npro, K)
-    p = Polynomial([-K^1.5*Pt, 0, K^1.5, 2Npro-Pt, 0, 1])
-    f = x -> isreal(x) && 0<=real(x)<=Pt
-    r = roots(p)
-    r = filter(f, r)
-    if isempty(r)
-        if Pt < 1e-3
-            return 0.0
-        else
-            return Pt
+    function root(Pt, Npro, K)
+        p = Polynomial([-K^1.5*Pt, 0, K^1.5, 2Npro-Pt, 0, 1])
+        f = x -> isreal(x) && 0<=real(x)<=Pt
+        r = roots(p)
+        r = filter(f, r)
+        if isempty(r)
+            if Pt < 1e-3
+                return 0.0
+            else
+                return Pt
+            end
         end
-    end
-    return real(first(r))^2
-end
-
-@register_symbolic root(Pt, Npro, K)
-
-
-# Nₒ and Nₜ are the numbers of repressilator and titration sponge plasmids (means of N̄ₒ and N̄ₜ)
-# P₁, P₂, P₃ are the three repressors (cI, lacI, tetR)
-# gfp is the green fluorescent protein reporter
-
-function jump_improved_repressilator()
-    r = @reaction_network begin
-        @parameters n K[1:3]
-        @species gr(t) V(t)
-        a*gr * N̄ₒ*V, ∅ --> Nₒ
-        a*gr * N̄ₜ*V, ∅ --> Nₜ
-        a*gr * (hillr(root(P₃/V, (Nₒ + Nₜ)/V, K[3]), λ * (Nₒ + Nₜ)/V, K[3], n))*V, ∅ --> b̄ * P₁
-        a*gr * (hillr(root(P₁/V, Nₒ/V, K[1]), λ * Nₒ/V, K[1], n))*V, ∅ --> b̄ * P₂
-        a*gr * (hillr(root(P₂/V, Nₒ/V, K[2]), λ * Nₒ/V, K[2], n))*V, ∅ --> b̄ * P₃
-        a*gr * (hillr(root(P₃/V, (Nₒ + Nₜ)/V, K[3]), λ * (Nₒ + Nₜ)/V, K[3], n))*V, ∅ --> b̄ * gfp
-        b*gr, Nₒ --> ∅ # Degradation is only mimicing dilution
-        b*gr, Nₜ --> ∅
-        b*gr, P₁ --> ∅
-        b*gr, P₂ --> ∅
-        b*gr, P₃ --> ∅
-        b*gr, gfp --> ∅
+        return real(first(r))^2
     end
 
-    u0_integers = [:P₁ => 160.0, :P₂ => 55.0, :Nₒ => 10.0, :Nₜ => 40.0, :P₃ => 2400.0, :gfp => 0.0, :gr => 1.0, :V => 1.0]
+    @register_symbolic root(Pt, Npro, K)
 
-    tspan = (0., 60.)
-    ps = [:K => K, :N̄ₒ => N̄ₒ, :N̄ₜ => N̄ₜ, :λ => λ, :n => n, :b̄ => b̄, :b => 0.0, :a => 1.0, :P₁_free => 50.0, :P₂_free => 40.0, :P₃_free => 30.0]
-    dprob = DiscreteProblem(r, u0_integers, tspan, ps)
-    jprob = JumpProblem(r, dprob, Direct(); save_positions = (false, false))
-    return jprob
-end
+    # Nₒ and Nₜ are the numbers of repressilator and titration sponge plasmids (means of N̄ₒ and N̄ₜ)
+    # P₁, P₂, P₃ are the three repressors (cI, lacI, tetR)
+    # gfp is the green fluorescent protein reporter
+    function sde_improved_repressilator()
+        r = @reaction_network begin
+            @parameters n K[1:3]
+            @species gr(t) V(t)
+            a*gr * N̄ₒ*V, ∅ --> Nₒ
+            a*gr * N̄ₜ*V, ∅ --> Nₜ
+            a*gr * (hillr(root(P₃/V, (Nₒ + Nₜ)/V, K[3]), λ * (Nₒ + Nₜ)/V, K[3], n))*V, ∅ --> b̄ * P₁
+            a*gr * (hillr(root(P₁/V, Nₒ/V, K[1]), λ * Nₒ/V, K[1], n))*V, ∅ --> b̄ * P₂
+            a*gr * (hillr(root(P₂/V, Nₒ/V, K[2]), λ * Nₒ/V, K[2], n))*V, ∅ --> b̄ * P₃
+            a*gr * (hillr(root(P₃/V, (Nₒ + Nₜ)/V, K[3]), λ * (Nₒ + Nₜ)/V, K[3], n))*V, ∅ --> b̄ * gfp
+            b*gr, Nₒ --> ∅ # Degradation is only mimicing dilution
+            b*gr, Nₜ --> ∅
+            b*gr, P₁ --> ∅
+            b*gr, P₂ --> ∅
+            b*gr, P₃ --> ∅
+            b*gr, gfp --> ∅
+        end
 
-function plot_jump()
-    jprob = jump_improved_repressilator()
-    sol = solve(jprob, SSAStepper(); progress=true)
-    display(Plots.plot(sol, vars=[:P₁, :P₂, :P₃], xlabel="Time", ylabel="Concentration", title="Repressilator Protein Dynamics (Gillespie)", linewidth=2))
-    display(Plots.plot(sol, vars=[:Nₒ, :Nₜ], xlabel="Time", ylabel="Concentration", title="Plasmid Concentrations (Gillespie)", linewidth=2))
-    Plots.plot(sol, vars=[:gfp], xlabel="Time", ylabel="Concentration", title="Repressilator Protein Dynamics (Gillespie)", linewidth=2)
-
-    sol = solve(jprob, SSAStepper(); progress=true)
-    display(Plots.plot!(sol, vars=[:gfp], label="gfp - Cell 2", linestyle=:dash, linewidth=2))
-end
-
-function sde_improved_repressilator()
-    r = @reaction_network begin
-        @parameters n K[1:3]
-        @species gr(t) V(t)
-        a*gr * N̄ₒ*V, ∅ --> Nₒ
-        a*gr * N̄ₜ*V, ∅ --> Nₜ
-        a*gr * (hillr(root(P₃/V, (Nₒ + Nₜ)/V, K[3]), λ * (Nₒ + Nₜ)/V, K[3], n))*V, ∅ --> b̄ * P₁
-        a*gr * (hillr(root(P₁/V, Nₒ/V, K[1]), λ * Nₒ/V, K[1], n))*V, ∅ --> b̄ * P₂
-        a*gr * (hillr(root(P₂/V, Nₒ/V, K[2]), λ * Nₒ/V, K[2], n))*V, ∅ --> b̄ * P₃
-        a*gr * (hillr(root(P₃/V, (Nₒ + Nₜ)/V, K[3]), λ * (Nₒ + Nₜ)/V, K[3], n))*V, ∅ --> b̄ * gfp
-        b*gr, Nₒ --> ∅ # Degradation is only mimicing dilution
-        b*gr, Nₜ --> ∅
-        b*gr, P₁ --> ∅
-        b*gr, P₂ --> ∅
-        b*gr, P₃ --> ∅
-        b*gr, gfp --> ∅
+        u0 = [:P₁ => 0.0, :P₂ => 0.0, :Nₒ => 10.0, :Nₜ => 40.0, :P₃ => 1000.0, :gfp => 0.0, :gr => 1.0, :V => 1.0]
+        tspan = (0., 30.)
+        ps = [:K => K, :N̄ₒ => N̄ₒ, :N̄ₜ => N̄ₜ, :λ => λ, :n => n, :b̄ => b̄, :a => 15.0, :b => 4.0]
+        sde = SDEProblem(r, u0, tspan, ps; structural_simplify = true)
+        return sde
     end
-
-    u0 = [:P₁ => 0.0, :P₂ => 0.0, :Nₒ => 10.0, :Nₜ => 40.0, :P₃ => 1000.0, :gfp => 0.0, :gr => 1.0, :V => 1.0]
-    tspan = (0., 30.)
-    ps = [:K => K, :N̄ₒ => N̄ₒ, :N̄ₜ => N̄ₜ, :λ => λ, :n => n, :b̄ => b̄, :a => 15.0, :b => 4.0]
-    sde = SDEProblem(r, u0, tspan, ps; structural_simplify = true)
-    return sde
-end
-
-function plot_sde()
-    sde = sde_improved_repressilator()
-    method = EM()
-    sol_sde = solve(sde, method, dt=0.01)
-
-    display(Plots.plot(sol_sde, vars=[:P₁, :P₂, :P₃], xlabel="Time", ylabel="Concentration", title="Repressilator Protein Dynamics (SDE)", linewidth=2))
-    display(Plots.plot(sol_sde, vars=[:Nₒ, :Nₜ], xlabel="Time", ylabel="Concentration", title="Plasmid Concentrations (SDE)", linewidth=2))
-    Plots.plot(sol_sde, vars=[:gfp], xlabel="Time", ylabel="Concentration", title="Repressilator Protein Dynamics (SDE)", linewidth=2)
-
-    sol_sde = solve(sde, method, dt=0.01)
-    display(Plots.plot!(sol_sde, vars=[:gfp], label="gfp - Cell 2", linestyle=:dash, linewidth=2))
 end
